% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESur1.R
\name{CESur1}
\alias{CESur1}
\title{CESur1}
\usage{
CESur1(x)
}
\arguments{
\item{x}{a data.frame containing GeneID and gene expression of a single patient}
}
\description{
CESur1 is used for predicting the survival rate of lung adenocarcinoma patient
}
