% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESur2.R
\name{CESur2}
\alias{CESur2}
\title{CESur2}
\usage{
CESur2(x)
}
\arguments{
\item{x}{a data.frame containing GeneID and gene expression of a single patient}
}
\description{
CESur2 is used for predicting the survival rate of stage IA lung adenocarcinoma patient
}
